+++ src/string/mod.rs
@@ -12,6 +12,7 @@ use core::{
 };
 
 use crate::vec::{OwnedVecStorage, Vec, VecInner, VecStorage, ViewVecStorage};
+use crate::CapacityError;
 
 mod drain;
 pub use drain::Drain;
@@ -24,17 +25,25 @@ pub use drain::Drain;
 #[derive(Debug)]
 pub enum FromUtf16Error {
     /// The capacity of the `String` is too small for the given operation.
-    Capacity,
+    Capacity(CapacityError),
     /// Error decoding UTF-16.
-    DecodeUtf16Error(DecodeUtf16Error),
+    DecodeUtf16(DecodeUtf16Error),
 }
 
 impl fmt::Display for FromUtf16Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::Capacity => "insufficient capacity".fmt(f),
-            Self::DecodeUtf16Error(e) => write!(f, "invalid UTF-16: {}", e),
+            Self::Capacity(err) => write!(f, "{err}"),
+            Self::DecodeUtf16(err) => write!(f, "invalid UTF-16: {err}"),
         }
+    }
+}
+
+impl core::error::Error for FromUtf16Error {}
+
+impl From<CapacityError> for FromUtf16Error {
+    fn from(e: CapacityError) -> Self {
+        Self::Capacity(e)
     }
 }
 
@@ -164,10 +173,10 @@ impl<const N: usize> String<N> {
         for c in char::decode_utf16(v.iter().cloned()) {
             match c {
                 Ok(c) => {
-                    s.push(c).map_err(|_| FromUtf16Error::Capacity)?;
+                    s.push(c).map_err(|_| CapacityError)?;
                 }
                 Err(err) => {
-                    return Err(FromUtf16Error::DecodeUtf16Error(err));
+                    return Err(FromUtf16Error::DecodeUtf16(err));
                 }
             }
         }
@@ -253,7 +262,7 @@ impl<const N: usize> String<N> {
     /// assert!(b.len() == 2);
     ///
     /// assert_eq!(&[b'a', b'b'], &b[..]);
-    /// # Ok::<(), ()>(())
+    /// # Ok::<(), heapless::CapacityError>(())
     /// ```
     #[inline]
     pub fn into_bytes(self) -> Vec<u8, N> {
@@ -360,7 +369,7 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     ///
     /// let _s = s.as_str();
     /// // s.push('c'); // <- cannot borrow `s` as mutable because it is also borrowed as immutable
-    /// # Ok::<(), ()>(())
+    /// # Ok::<(), heapless::CapacityError>(())
     /// ```
     #[inline]
     pub fn as_str(&self) -> &str {
@@ -379,7 +388,7 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     /// let mut s: String<4> = String::try_from("ab")?;
     /// let s = s.as_mut_str();
     /// s.make_ascii_uppercase();
-    /// # Ok::<(), ()>(())
+    /// # Ok::<(), heapless::CapacityError>(())
     /// ```
     #[inline]
     pub fn as_mut_str(&mut self) -> &mut str {
@@ -411,7 +420,7 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     ///     vec.reverse();
     /// }
     /// assert_eq!(s, "olleh");
-    /// # Ok::<(), ()>(())
+    /// # Ok::<(), heapless::CapacityError>(())
     /// ```
     pub unsafe fn as_mut_vec(&mut self) -> &mut VecInner<u8, S> {
         &mut self.vec
@@ -433,11 +442,10 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     /// assert_eq!("foobar", s);
     ///
     /// assert!(s.push_str("tender").is_err());
-    /// # Ok::<(), ()>(())
+    /// # Ok::<(), heapless::CapacityError>(())
     /// ```
     #[inline]
-    #[allow(clippy::result_unit_err)]
-    pub fn push_str(&mut self, string: &str) -> Result<(), ()> {
+    pub fn push_str(&mut self, string: &str) -> Result<(), CapacityError> {
         self.vec.extend_from_slice(string.as_bytes())
     }
 
@@ -476,13 +484,12 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     /// assert!("abc123" == s.as_str());
     ///
     /// assert_eq!("abc123", s);
-    /// # Ok::<(), ()>(())
+    /// # Ok::<(), heapless::CapacityError>(())
     /// ```
     #[inline]
-    #[allow(clippy::result_unit_err)]
-    pub fn push(&mut self, c: char) -> Result<(), ()> {
+    pub fn push(&mut self, c: char) -> Result<(), CapacityError> {
         match c.len_utf8() {
-            1 => self.vec.push(c as u8).map_err(|_| {}),
+            1 => self.vec.push(c as u8).map_err(|_| CapacityError),
             _ => self
                 .vec
                 .extend_from_slice(c.encode_utf8(&mut [0; 4]).as_bytes()),
@@ -513,7 +520,7 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     /// s.truncate(2);
     ///
     /// assert_eq!("he", s);
-    /// # Ok::<(), ()>(())
+    /// # Ok::<(), heapless::CapacityError>(())
     /// ```
     #[inline]
     pub fn truncate(&mut self, new_len: usize) {
@@ -541,7 +548,7 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     /// assert_eq!(s.pop(), Some('f'));
     ///
     /// assert_eq!(s.pop(), None);
-    /// Ok::<(), ()>(())
+    /// Ok::<(), heapless::CapacityError>(())
     /// ```
     pub fn pop(&mut self) -> Option<char> {
         let ch = self.chars().next_back()?;
@@ -615,7 +622,7 @@ impl<S: VecStorage<u8> + ?Sized> StringInner<S> {
     /// assert!(s.is_empty());
     /// assert_eq!(0, s.len());
     /// assert_eq!(8, s.capacity());
-    /// Ok::<(), ()>(())
+    /// Ok::<(), heapless::CapacityError>(())
     /// ```
     #[inline]
     pub fn clear(&mut self) {
@@ -630,7 +637,8 @@ impl<const N: usize> Default for String<N> {
 }
 
 impl<'a, const N: usize> TryFrom<&'a str> for String<N> {
-    type Error = ();
+    type Error = CapacityError;
+
     fn try_from(s: &'a str) -> Result<Self, Self::Error> {
         let mut new = Self::new();
         new.push_str(s)?;
@@ -639,7 +647,7 @@ impl<'a, const N: usize> TryFrom<&'a str> for String<N> {
 }
 
 impl<const N: usize> str::FromStr for String<N> {
-    type Err = ();
+    type Err = CapacityError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let mut new = Self::new();
@@ -912,7 +920,7 @@ impl_try_from_num!(u64, 20);
 
 #[cfg(test)]
 mod tests {
-    use crate::{String, Vec};
+    use crate::{CapacityError, String, Vec};
 
     #[test]
     fn static_new() {
@@ -980,7 +988,7 @@ mod tests {
         assert!(s.len() == 3);
         assert_eq!(s, "123");
 
-        let _: () = String::<2>::try_from("123").unwrap_err();
+        let _: CapacityError = String::<2>::try_from("123").unwrap_err();
     }
 
     #[test]
@@ -991,7 +999,7 @@ mod tests {
         assert!(s.len() == 3);
         assert_eq!(s, "123");
 
-        let _: () = String::<2>::from_str("123").unwrap_err();
+        let _: CapacityError = String::<2>::from_str("123").unwrap_err();
     }
 
     #[test]